requirements.txt
numpy
pyaudio
onnxruntime
luma.oled
gpiozero
